pyproject.toml
[tool.poetry]
name = "xenon-bot"
version = "0.4.0"
description = "An elegant QQ Bot using Mirai and Graia Project."
authors = ["McZoo Dev Team"]
license = "AGPL-v3"

[tool.poetry.dependencies]
python = "3.9"
prompt-toolkit = "^3.0.19"
types-PyYAML = "^5.4.6"
graia-application-mirai = "0.19.2"
graia-broadcast = "0.8.11"
graia-scheduler = "0.0.4"
graia-saya = "0.0.9"
aiosqlite = "^0.17.0"
croniter = "^0.3.37"
loguru = "^0.5.3"
graiax-v4-dependency = "^1.0.1"

[tool.poetry.dev-dependencies]
mcstatus = "^6.5.0"
Sphinx = "^4.1.2"
sphinx-rtd-theme = "^1.0.0rc1"
recommonmark = "^0.7.1"
autodoc = "^0.5.0"
autodoc-pydantic = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
